refactor: Adicionando segurança ao service de user.

+++ back/Ordem-De-Servico/src/main/java/com/ordemDeServico/service/UserService.java
@@ -1,13 +1,17 @@
 package com.ordemDeServico.service;
 
 import com.ordemDeServico.exceptions.EmailRegisteredException;
+import com.ordemDeServico.exceptions.NotExistException;
+import com.ordemDeServico.exceptions.UnauthorizedAccessException;
 import com.ordemDeServico.repository.UsuarioRepository;
 import com.ordemDeServico.service.event.UsuarioCriado;
 import com.ordemDeServico.model.Usuario;
 import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.rmi.AccessException;
 import java.util.List;
 import java.util.Optional;
 
@@ -34,7 +38,7 @@ public class UserService {
 
     public Usuario buscarPorId(int id) {
         return usuarioRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
+                .orElseThrow(() -> new NotExistException("Usuário não encontrado"));
     }
 
     public List<Usuario> listarTodos() {
@@ -43,14 +47,25 @@ public class UserService {
 
     public Usuario atualizar(int id, Usuario dados) {
         Usuario usuario = buscarPorId(id);
+        Usuario logado = (Usuario) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!logado.getUsername().equals(usuario.getUsername())) {
+            throw new UnauthorizedAccessException("Acesso Negado.");
+        }
+
+        String encryptedPassword = new BCryptPasswordEncoder().encode(dados.getPassword());
         usuario.setNome(dados.getNome());
         usuario.setEmail(dados.getEmail());
         usuario.setSenha(dados.getSenha());
+        usuario.setSenha(encryptedPassword);
         return usuarioRepository.save(usuario);
     }
 
     public void deletar(int  id) {
         Usuario usuario = buscarPorId(id);
+        Usuario logado = (Usuario) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!logado.getUsername().equals(usuario.getUsername())) {
+            throw new UnauthorizedAccessException("Acesso Negado.");
+        }
         usuarioRepository.delete(usuario);
     }
